Number day 1 videos by paralleling the exercise
The exercise is listed by "exercise_title:" followed by "exercise:". In contrast, events had just been listed by "name:" within the broader header of "events:". I am listing the events by "event_title:" followed by "event:" to see if this causes the numbering to work for events as it does for the exercise.