Remove col-md-4 in sidebar
Within the main page setups, the sidebar is given a col-md-4, which means it takes up 4 of 12 possible columns. That happens outside of sidebar.html. Then, sidebar.html is called. For the locations sidebar in this file, there was no further columning command. For the other sidebars, this file had given them an additional col-md-4. I believe this caused those sidebars to be squeezed into 1/3 of the horizontal space allocated. In other words, the links in the sidebar were squeezed into 1/3 of the allocated space, which was already 1/3 of the width of the screen since 1/3 of the width of the screen had been allocated to the sidebar. Because the links looked correct in the locations sidebar and this code did not have the extra col-md-4, I think that removing the redundant col-md-4 commands in this file will make the links take up the full sidebar. This commit comments out those lines, so that I can re-activate them quickly if this causes a problem.