fixed faculty description
If I'm not mistaken, you are not supposed to create a "people.yml" file but rather insert all the necessary information in the particular sub-yml files. Fixed that for you. Plus, I removed the quotation marks around Jae Yeon Kims book as they, for me, sometimes lead to problems. Hope that helps.