Update faculty.yml
I removed the colon (:) from the bio of Jedrzej Wydra since it caused a build error on CircleCI. The error prevented the main site from building successfully. The reason the colon caused the error is that the builder read specialty: as a key/mapping instead of normal text. Since it is illegal to have keys as part of values, it caused the build error.