Implement basic searching
This adds basic lunr-based search to the videos. The video page embeds a
LARGE JSON object that's generated from the existing video collection
(using the jsonify liquid filter).

JavaScript uses this to build a lunr index of all the video data. The
search itself performs an unweighted lunr search against title,
content, and author fields (more can be added if needed).

When any search is performed, the video listing container gets a class
that indicates it's in "searching" mode; when in searching mode, all
videos are hidden by default. If lunr finds any results, then the
corresponding video card is given a class that causes CSS to display it.

If no results are found, then no videos will display at all; a future
commit will add a "no results" message to indicate that things aren't
broken.

A few drawbacks to this approach so far:

- requires all videos to be actually displayed on the current page;
  if/when pagination is added, then we will need to deal with that.
- the entire JSON blob from jekyll is dumped directly into the page. At
  a minimum it would be good to reduce this to just the necessary data
  needed by lunr, but really we should build this index when building the
  site, and fetch it via ajax the same way people are fetched. Since the
  video collection lives as markdown pages, we can't do this exactly the
  same way as with people searching, but it will be a similar idea.

Both of these can be addressed by iteration.