Generate smaller JS data for lunr
Rather than generate a massive JS object with *all* the Jekyll data,
generate a smaller object with *just* the data that lunr cares about (as
defined in config.yml).

It would still be better to pregenerate a lunr index when building the
site, but this at least reduces the page payload significantly by
eliminating useless data.