Add (nonfunctional, currently) sidebar filters
Add some filter configuration to the main site config file that
determines which filters to pluck from the video files, and how to
display them.

Add a loop/template that generates a list of filters based on actual
data used in the video files, and displays them as possible filters.

This does not yet include actual filtering functionality.