Connect filters to lunr
This includes... some copypasta from the existing people search, to
duplicate the logic. Further refactoring would DRY it up a bit and maybe
make a generic multi-search pattern. The data source and UI are slightly
different between people and video so it would take some care to make
that happen.