update people.yml
Hi, there was a syntax error in this file that caused me to be unable to deploy a hotfix. I added forward slashes in the book title on line 18. You won't see these on the website, but it is required syntax when you use double quotes inside of a string.